src: lock-free init_process_flags
Fix https://github.com/nodejs/node/issues/45152

PR-URL: https://github.com/nodejs/node/pull/45221

+++ src/node.cc
@@ -429,7 +429,7 @@ void ResetSignalHandlers() {
 #endif  // __POSIX__
 }
 
-static std::atomic<uint64_t> init_process_flags = 0;
+static std::atomic<uint32_t> init_process_flags = 0;
 
 static void PlatformInit(ProcessInitializationFlags::Flags flags) {
   // init_process_flags is accessed in ResetStdio(),
